refactor: move the console client under the correct package
